Implement footer (#22)
closes #16 

* Introduce footer

* Adjust styling for footer

Looks like the layout design didn't count with a potential page footer, this fix should provide at least somewhat improved design for current pages

* Revert "Adjust styling for footer"

This reverts commit 7718d06d776b76a1bcbe5503af0915ae59cb616f.

* Add footer visibility

The footer is not required on all the pages using the layout. This commit adds an ability to hide the footer element using the ViewBag in specific Views. Default visibility is set to true.

* Fix page styling

Co-authored-by: Filip Dorian Pindej <fpindej@users.noreply.github.com>